+++ interface/src/routes/+layout.svelte
@@ -38,10 +38,7 @@
 		socket.on('close', handleClose);
 		socket.on('error', handleError);
 		socket.on('rssi', handleNetworkStatus);
-		socket.on('infoToast', handleInfoToast);
-		socket.on('successToast', handleSuccessToast);
-		socket.on('warningToast', handleWarningToast);
-		socket.on('errorToast', handleErrorToast);
+		socket.on('notification', handleNotification);
 		if ($page.data.features.analytics) socket.on('analytics', handleAnalytics);
 		if ($page.data.features.battery) socket.on('battery', handleBattery);
 		if ($page.data.features.download_firmware) socket.on('otastatus', handleOAT);
@@ -52,10 +49,7 @@
 		socket.off('open', handleOpen);
 		socket.off('close', handleClose);
 		socket.off('rssi', handleNetworkStatus);
-		socket.off('infoToast', handleInfoToast);
-		socket.off('successToast', handleSuccessToast);
-		socket.off('warningToast', handleWarningToast);
-		socket.off('errorToast', handleErrorToast);
+		socket.off('notification', handleNotification);
 		socket.off('battery', handleBattery);
 		socket.off('otastatus', handleOAT);
 	};
@@ -88,10 +82,24 @@
 
 	const handleError = (data: any) => console.error(data);
 
-	const handleInfoToast = (data: string) => notifications.info(data, 5000);
-	const handleWarningToast = (data: string) => notifications.warning(data, 5000);
-	const handleErrorToast = (data: string) => notifications.error(data, 5000);
-	const handleSuccessToast = (data: string) => notifications.success(data, 5000);
+	const handleNotification = (data: any) => {
+		switch (data.type) {
+			case 'info':
+				notifications.info(data.message, 5000);
+				break;
+			case 'warning':
+				notifications.warning(data.message, 5000);
+				break;
+			case 'error':
+				notifications.error(data.message, 5000);
+				break;
+			case 'success':
+				notifications.success(data.message, 5000);
+				break;
+			default:
+				break;
+		}
+	};
 
 	const handleAnalytics = (data: Analytics) => analytics.addData(data);