+++ lib/framework/EventSocket.h
@@ -25,7 +25,7 @@
 #define EVENT_SERVICE_PATH "/ws/events"
 
 typedef std::function<void(JsonObject &root, int originId)> EventCallback;
-typedef std::function<void(const String &originId, bool sync)> SubscribeCallback;
+typedef std::function<void(const String &originId)> SubscribeCallback;
 
 class EventSocket
 {
@@ -40,14 +40,8 @@ public:
 
   void onSubscribe(String event, SubscribeCallback callback);
 
-  void emit(String event, String payload);
-
-  void emit(const char *event, const char *payload);
-
-  void emit(const char *event, const char *payload, const char *originId, bool onlyToSameOrigin = false);
-  // if onlyToSameOrigin == true, the message will be sent to the originId only, otherwise it will be broadcasted to all clients except the originId
-
   void emitEvent(String event, JsonObject &jsonObject, const char *originId = "", bool onlyToSameOrigin = false);
+  // if onlyToSameOrigin == true, the message will be sent to the originId only, otherwise it will be broadcasted to all clients except the originId
 
 private:
   PsychicHttpServer *_server;